为class添加爱了成绩比例字段； 完成了签到功能； 完成了report和签到的成绩功能，可能JPA有点小bug。
还需要：
1、授权
2、对抗练习
3、redis进行时限控制
4、邮箱，用邮箱来进行账号激活？？？？？

+++ src/main/java/com/susu/se/controller/Report/AppraiseReportController.java
@@ -18,12 +18,12 @@ public class AppraiseReportController {
     private AppraiseReportService appraiseReportService;
 
     @PostMapping("/appraisebyta")
-    public Result<String> appraiseByTA(@RequestParam("reportid") Integer reportId,@RequestParam("score") Integer score,@RequestParam("pingyu") String pingYu, @RequestParam("assistantid") Integer assistantId){
+    public Result<String> appraiseByTA(@RequestParam("reportid") Integer reportId,@RequestParam("score") Double score,@RequestParam("pingyu") String pingYu, @RequestParam("assistantid") Integer assistantId){
         return appraiseReportService.appraiseReportByTA(reportId, score, pingYu, assistantId);
     }
 
     @PostMapping("/appraisebyte")
-    public Result<String> appraiseByTeacher(@RequestParam("reportid") Integer reportId,@RequestParam("score") Integer score,@RequestParam("pingyu") String pingYu, @RequestParam("teacherid") Integer teacherId){
+    public Result<String> appraiseByTeacher(@RequestParam("reportid") Integer reportId,@RequestParam("score") Double score,@RequestParam("pingyu") String pingYu, @RequestParam("teacherid") Integer teacherId){
         return appraiseReportService.appraiseReportByTeacher(reportId, score, pingYu, teacherId);
     }
